hexagon: factor out hvx-sqrt.h
